Move test helpers to new module
