+++ ComposeGallery/build.gradle
@@ -14,7 +14,6 @@ android {
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
-
     }
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
@@ -38,26 +37,29 @@ android {
 
 dependencies {
     implementation kotlinDependencies.kotlinStdlib
+
     implementation androidxDependencies.ktxCore
     implementation androidxDependencies.appCompat
     implementation androidxDependencies.window
+
     implementation composeDependencies.composeUI
     implementation composeDependencies.composeRuntime
     implementation composeDependencies.composeMaterial
     implementation composeDependencies.composeUITooling
     implementation composeDependencies.activityCompose
+
     implementation microsoftDependencies.twoPaneLayout
 
     implementation googleDependencies.material
 
-    androidTestImplementation(testDependencies.androidxTestCore)
-    androidTestImplementation(testDependencies.androidxTestRules)
-    androidTestImplementation(testDependencies.androidxTestRunner)
-    androidTestImplementation(testDependencies.composeUITest)
-    androidTestImplementation(testDependencies.composeJunit)
-    androidTestImplementation(testDependencies.uiAutomator)
-    androidTestImplementation(testDependencies.windowTest)
+    androidTestImplementation testDependencies.androidxTestCore
+    androidTestImplementation testDependencies.androidxTestRules
+    androidTestImplementation testDependencies.androidxTestRunner
+    androidTestImplementation testDependencies.espressoCore
+    androidTestImplementation testDependencies.composeUITest
+    androidTestImplementation testDependencies.composeJunit
+    androidTestImplementation testDependencies.windowTest
+    androidTestImplementation project(':foldable-test')
 
-    debugImplementation(testDependencies.composeUITestManifest)
-
+    debugImplementation testDependencies.composeUITestManifest
 }
