+++ ComposeGallery/src/androidTest/java/com/microsoft/device/display/samples/composegallery/TopAppBarTest.kt
@@ -20,6 +20,7 @@ import com.microsoft.device.display.samples.composegallery.ui.ComposeGalleryThem
 import com.microsoft.device.display.samples.composegallery.ui.view.ComposeGalleryApp
 import com.microsoft.device.display.samples.composegallery.ui.view.DetailPane
 import com.microsoft.device.display.samples.composegallery.ui.view.ListPane
+import com.microsoft.device.dualscreen.foldable_test.getString
 import org.junit.Rule
 import org.junit.Test
