+++ foldable-test/src/main/java/com/microsoft/device/dualscreen/foldable_test/DeviceModel.kt
